refactor(uat): extract cells_io.py as the single cells.jsonl+sidecar codec (w1)
The cells.jsonl schema lived in two places with no shared definition: the
writer in orchestrator._write_cells_jsonl and the reader (JSONL parsing +
accounting sidecar read) inline in _cli.py's _handle_report. Extract both
into tests/uat/cells_io.py so the orchestrator (writer) and `make
uat-report` (reader) consume one schema. Sidecar filename convention
(cells_accounting_path) is also unified. On-disk content is byte-identical;
only the owning code moved. Part of uat-execute-path-unification w1.
